Made some button changes
Details are only stored when the buttons aren't red but clicking a
button will change the colour so it can be booked again. You can no
longer unbook bronze seats. Gold seats now have a 1 in 10 chance of
winning a backstage pass and silver seats are now informed they are
entitled to a free programme. These details will stil need to be
reflected in the array though.